Split 'get unfiltered projects' out to ProjectsGenerator
Also, only define TestProject etc in the tests which ask for it
