+++ lib/kennel.rb
@@ -9,6 +9,7 @@ require "kennel/compatibility"
 require "kennel/utils"
 require "kennel/progress"
 require "kennel/filter"
+require "kennel/projects_generator"
 require "kennel/syncer"
 require "kennel/id_map"
 require "kennel/api"
@@ -126,12 +127,14 @@ module Kennel
       @api ||= Api.new(ENV.fetch("DATADOG_APP_KEY"), ENV.fetch("DATADOG_API_KEY"))
     end
 
+    def projects_generator
+      @projects_generator ||= ProjectsGenerator.new
+    end
+
     def generated
       @generated ||= begin
         Progress.progress "Generating" do
-          load_all
-
-          projects = Models::Project.recursive_subclasses.map(&:new)
+          projects = projects_generator.projects
           Kennel::Filter.filter_resources!(projects, :kennel_id, filter.project_filter, "projects", "PROJECT")
 
           parts = Utils.parallel(projects, &:validated_parts).flatten(1)
@@ -151,48 +154,6 @@ module Kennel
           parts
         end
       end
-    end
-
-    def load_all
-      # load_all's purpose is to "require" all the .rb files under './projects',
-      # also with reference to ./teams and ./parts. What happens if you call it
-      # more than once?
-      #
-      # For a reason yet to be investigated, Zeitwerk rejects second and subsequent calls.
-      # But even if we skip over the Zeitwerk part, the nature of 'require' is
-      # somewhat one-way: we're not providing any mechanism to *un*load things.
-      # As long as the contents of `./projects`, `./teams` and `./parts` doesn't
-      # change between calls, then simply by no-op'ing subsequent calls to `load_all`
-      # we can have `load_all` appear to be idempotent.
-      loader = Zeitwerk::Loader.new
-      Dir.exist?("teams") && loader.push_dir("teams", namespace: Teams)
-      Dir.exist?("parts") && loader.push_dir("parts")
-      loader.setup
-      loader.eager_load # TODO: this should not be needed but we see hanging CI processes when it's not added
-
-      # TODO: also auto-load projects and update expected path too
-      ["projects"].each do |folder|
-        Dir["#{folder}/**/*.rb"].sort.each { |f| require "./#{f}" }
-      end
-    rescue NameError => e
-      message = e.message
-      raise unless klass = message[/uninitialized constant (.*)/, 1]
-
-      # inverse of zeitwerk lib/zeitwerk/inflector.rb
-      path = klass.gsub("::", "/").gsub(/([a-z])([A-Z])/, "\\1_\\2").downcase + ".rb"
-      expected_path = (path.start_with?("teams/") ? path : "parts/#{path}")
-
-      # TODO: prefer to raise a new exception with the old backtrace attacked
-      e.define_singleton_method(:message) do
-        "\n" + <<~MSG.gsub(/^/, "  ")
-          #{message}
-          Unable to load #{klass} from #{expected_path}
-          - Option 1: rename the constant or the file it lives in, to make them match
-          - Option 2: Use `require` or `require_relative` to load the constant
-        MSG
-      end
-
-      raise
     end
   end
 end
