+++ test/kennel/models/base_test.rb
@@ -4,6 +4,8 @@ require_relative "../../test_helper"
 SingleCov.covered!
 
 describe Kennel::Models::Base do
+  with_test_classes
+
   class TestBase < Kennel::Models::Base
     settings :foo, :bar, :unset
     defaults(
