+++ test/kennel/models/dashboard_test.rb
@@ -4,6 +4,8 @@ require_relative "../../test_helper"
 SingleCov.covered!
 
 describe Kennel::Models::Dashboard do
+  with_test_classes
+
   class TestDashboard < Kennel::Models::Dashboard
   end
