+++ test/kennel/models/monitor_test.rb
@@ -4,6 +4,8 @@ require_relative "../../test_helper"
 SingleCov.covered!
 
 describe Kennel::Models::Monitor do
+  with_test_classes
+
   class TestMonitor < Kennel::Models::Monitor
   end
