+++ test/kennel/models/project_test.rb
@@ -4,6 +4,8 @@ require_relative "../../test_helper"
 SingleCov.covered!
 
 describe Kennel::Models::Project do
+  with_test_classes
+
   describe ".file_location" do
     let(:plain_project_class) do
       Class.new(Kennel::Models::Project) do
