+++ test/kennel/models/record_test.rb
@@ -4,6 +4,8 @@ require_relative "../../test_helper"
 SingleCov.covered!
 
 describe Kennel::Models::Record do
+  with_test_classes
+
   class TestRecord < Kennel::Models::Record
     settings :foo, :bar, :override, :unset
     defaults(
