+++ test/kennel/models/slo_test.rb
@@ -4,6 +4,8 @@ require_relative "../../test_helper"
 SingleCov.covered!
 
 describe Kennel::Models::Slo do
+  with_test_classes
+
   class TestSlo < Kennel::Models::Slo
   end
