+++ test/kennel/subclass_tracking_test.rb
@@ -4,6 +4,8 @@ require_relative "../test_helper"
 SingleCov.covered!
 
 describe Kennel::SubclassTracking do
+  with_test_classes
+
   describe ".recursive_subclasses" do
     it "registers all created projects and subclasses" do
       Kennel::Models::Project.recursive_subclasses.must_equal [TestProject, SubTestProject]
